+++ liquid/golden/if_tag.py
@@ -274,6 +274,7 @@ cases = [
         expect="2",
         error=False,
         strict=True,
+        future=True,
     ),
     Case(
         description="extra else blocks are ignored",
@@ -282,6 +283,7 @@ cases = [
         expect="2",
         error=False,
         strict=True,
+        future=True,
     ),
     Case(
         description="extra elsif blocks are ignored",
@@ -290,6 +292,7 @@ cases = [
         expect="2",
         error=False,
         strict=True,
+        future=True,
     ),
     Case(
         description="empty array equals special empty",